requirements.txt
PyYAML==5.3
Pillow==7.1.2
pandas==0.25.1
torch==1.5.0
torchvision==0.6.0
albumentations==0.4.5
pytorch_lightning==0.7.1
